:bug: Fix average reviews score not calculating properly

+++ src/reviews/reviews.service.ts
@@ -1,17 +1,26 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ConflictException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Recipe } from 'src/recipes/recipes.entity';
 import { Review } from 'src/reviews/reviews.entity';
 import { User } from '../users/users.entity';
+import { CreateReviewDto } from './dto/create-review.dto';
+import { JwtPayload } from '../auth/strategies/jwt/jwt-payload.interface';
+import { RecipesService } from '../recipes/recipes.service';
+import { UsersService } from '../users/users.service';
+import { UpdateReviewDto } from './dto/update-review.dto';
 
 @Injectable()
 export class ReviewsService {
   constructor(
     @InjectRepository(Review)
     private readonly reviewsRepository: Repository<Review>,
-    @InjectRepository(Recipe)
-    private readonly recipesRepository: Repository<Recipe>,
+    private readonly recipesService: RecipesService,
+    private readonly usersService: UsersService,
   ) {}
 
   async findByRecipe(recipe: Recipe): Promise<Review[]> {
@@ -43,13 +52,38 @@ export class ReviewsService {
    * @returns {Promise<Recipe>} Promise of the reviewed recipe.
    * @memberof ReviewsService
    */
-  async create(review: Review): Promise<Review> {
-    review.recipe.averageReviewScore = this.calculateReviewsAverage(
-      review.recipe,
-    );
-    this.recipesRepository.save(review.recipe);
+  async create(
+    currentUser: JwtPayload,
+    recipeSlug: string,
+    createReviewDto: CreateReviewDto,
+  ): Promise<Review> {
+    const recipe = await this.recipesService.findOneBySlug(recipeSlug);
+    if (!recipe) {
+      throw new NotFoundException({ errors: { slug: 'Recipe not found' } });
+    }
 
-    return this.reviewsRepository.save(review);
+    const author = await this.usersService.findOneByEmail(currentUser.email);
+    let review = await this.reviewsRepository.findOne({
+      where: {
+        author,
+        recipe,
+      },
+    });
+    if (review) {
+      throw new ConflictException({
+        errors: { author: 'Recipe has already been reviewed' },
+      });
+    }
+
+    review = await this.reviewsRepository.save(
+      new Review({ ...createReviewDto, author }),
+    );
+    recipe.reviews.push(review);
+    recipe.averageReviewScore = this.calculateReviewsAverage(recipe);
+
+    await this.recipesService.update(recipe);
+
+    return review;
   }
 
   /**
@@ -60,13 +94,39 @@ export class ReviewsService {
    * @returns {Promise<Recipe>} Promise of the reviewed recipe.
    * @memberof ReviewsService
    */
-  async update(review: Review): Promise<Review> {
-    review.recipe.averageReviewScore = this.calculateReviewsAverage(
-      review.recipe,
-    );
-    this.recipesRepository.save(review.recipe);
+  async update(
+    currentUser: JwtPayload,
+    recipeSlug: string,
+    updateReviewDto: UpdateReviewDto,
+  ): Promise<Review> {
+    let recipe = await this.recipesService.findOneBySlug(recipeSlug);
+    if (!recipe) {
+      throw new NotFoundException({ errors: { slug: 'Recipe not found' } });
+    }
 
-    return this.reviewsRepository.save(review);
+    const author = await this.usersService.findOneByEmail(currentUser.email);
+    let review = await this.reviewsRepository.findOne({
+      where: {
+        author,
+        recipe,
+      },
+    });
+    if (!review) {
+      throw new ConflictException({
+        errors: { author: 'No review has been left for this recipe' },
+      });
+    }
+
+    review.text = updateReviewDto.text;
+    review.score = updateReviewDto.score;
+
+    review = await this.reviewsRepository.save(review);
+
+    recipe = await this.recipesService.findOneBySlug(recipeSlug);
+    recipe.averageReviewScore = this.calculateReviewsAverage(recipe);
+    await this.recipesService.update(recipe);
+
+    return review;
   }
 
   /**
@@ -77,13 +137,32 @@ export class ReviewsService {
    * @returns {Promise<Recipe>} Promise of the reviewed recipe.
    * @memberof ReviewsService
    */
-  async delete(review: Review): Promise<Review> {
-    review.recipe.averageReviewScore = this.calculateReviewsAverage(
-      review.recipe,
-    );
-    this.recipesRepository.save(review.recipe);
+  async delete(currentUser: JwtPayload, recipeSlug: string): Promise<Review> {
+    let recipe = await this.recipesService.findOneBySlug(recipeSlug);
+    if (!recipe) {
+      throw new NotFoundException({ errors: { slug: 'Recipe not found' } });
+    }
 
-    return this.reviewsRepository.remove(review);
+    const author = await this.usersService.findOneByEmail(currentUser.email);
+    let review = await this.reviewsRepository.findOne({
+      where: {
+        author,
+        recipe,
+      },
+    });
+    if (!review) {
+      throw new ConflictException({
+        errors: { author: 'No review has been left for this recipe' },
+      });
+    }
+
+    review = await this.reviewsRepository.remove(review);
+
+    recipe = await this.recipesService.findOneBySlug(recipeSlug);
+    recipe.averageReviewScore = this.calculateReviewsAverage(recipe);
+    await this.recipesService.update(recipe);
+
+    return review;
   }
 
   /**
@@ -95,13 +174,13 @@ export class ReviewsService {
    * @memberof ReviewsService
    */
   private calculateReviewsAverage(recipe: Recipe): number {
-    if (!recipe.reviews) {
-      return;
+    if (!recipe.reviews || !recipe.reviews.length) {
+      return 0;
     }
 
     return (
       recipe.reviews.reduce((avg, review) => avg + review.score, 0) /
-        recipe.reviews.length || 0
+      recipe.reviews.length
     );
   }
 }
